BIG REFACTOR Move poa validation, json schema validation and building of veteran model up to a parent class
Make testing easier via poa verification stub

Add poa validation to v1 526 posts
