+++ modules/claims_api/app/controllers/claims_api/v0/claims_controller.rb
@@ -7,6 +7,7 @@ module ClaimsApi
   module V0
     class ClaimsController < ApplicationController
       skip_before_action(:authenticate)
+      skip_before_action(:verify_power_of_attorney)
 
       def index
         claims = service.all
