+++ modules/claims_api/app/controllers/claims_api/v1/claims_controller.rb
@@ -7,7 +7,6 @@ module ClaimsApi
   module V1
     class ClaimsController < ApplicationController
       before_action { permit_scopes %w[claim.read] }
-      before_action :verify_power_of_attorney, if: :poa_request?
 
       def index
         claims = service.all
@@ -35,25 +34,6 @@ module ClaimsApi
 
       def service
         ClaimsApi::UnsynchronizedEVSSClaimService.new(target_veteran)
-      end
-
-      def target_veteran
-        if poa_request?
-          ClaimsApi::Veteran.from_headers(request.headers)
-        else
-          ClaimsApi::Veteran.from_identity(identity: @current_user)
-        end
-      end
-
-      def verify_power_of_attorney
-        verifier = EVSS::PowerOfAttorneyVerifier.new(target_veteran)
-        verifier.verify(@current_user)
-      end
-
-      def poa_request?
-        # if any of the required headers are present we should attempt to use headers
-        headers_to_check = ['HTTP_X_VA_SSN', 'HTTP_X_VA_Consumer-Username', 'HTTP_X_VA_Birth_Date']
-        (request.headers.to_h.keys & headers_to_check).length.positive?
       end
     end
   end
