+++ modules/claims_api/spec/requests/v1/disability_compensation_request_spec.rb
@@ -14,6 +14,7 @@ RSpec.describe 'Disability Claims ', type: :request do
       'X-VA-Gender': 'M' }
   end
   let(:scopes) { %w[claim.write] }
+  before(:each) { stub_poa_verification }
 
   describe '#526' do
     let(:data) { File.read(Rails.root.join('modules', 'claims_api', 'spec', 'fixtures', 'form_526_json_api.json')) }
